# Defer loading of umap to speed up import