--- a/.travis.yml
+++ b/.travis.yml
@@ -3,6 +3,7 @@ language: python
 python:
   - '3.6'
   - '3.7'
+  - '3.8-dev'
 cache: pip
 install:
   - pip install docutils sphinx