--- a/.travis.yml
+++ b/.travis.yml
@@ -3,7 +3,7 @@ language: python
 python:
   - '3.6'
   - '3.7'
-  - '3.8-dev'
+  #- '3.8-dev' # https://github.com/numpy/numpy/issues/13790
 cache: pip
 install:
   - pip install docutils sphinx