--- a/requirements.txt
+++ b/requirements.txt
@@ -7,7 +7,7 @@ seaborn
 h5py
 tables
 tqdm
-importlib_metadata
+importlib_metadata ; python_version<'3.8'
 # exclude buggy versions
 scikit-learn >= 0.19.1, != 0.21.0, != 0.21.1
 statsmodels>=0.10.0rc2