--- a/scanpy/utils.py
+++ b/scanpy/utils.py
@@ -25,29 +25,29 @@ EPS = 1e-15
 
 def check_versions():
     from distutils.version import LooseVersion
+    from importlib_metadata import version
 
     if sys.version_info < (3, 6):
         warnings.warn('Scanpy prefers Python 3.6 or higher. '
                       'Currently, Python 3.5 leads to a bug in `tl.marker_gene_overlap` '
                       'and we might stop supporting it in the future.')
 
-    import anndata, umap
-    # NOTE: pytest does not correctly retrieve anndata's version? why?
-    #       use the following hack...
-    if anndata.__version__ != '0+unknown':
-        if anndata.__version__ < LooseVersion('0.6.10'):
-            from . import __version__
-            raise ImportError('Scanpy {} needs anndata version >=0.6.10, not {}.\n'
-                              'Run `pip install anndata -U --no-deps`.'
-                              .format(__version__, anndata.__version__))
+    anndata_version = version("anndata")
+    umap_version = version("umap-learn")
 
-    if umap.__version__ < LooseVersion('0.3.0'):
+    if anndata_version < LooseVersion('0.6.10'):
+        from . import __version__
+        raise ImportError('Scanpy {} needs anndata version >=0.6.10, not {}.\n'
+                            'Run `pip install anndata -U --no-deps`.'
+                            .format(__version__, anndata_version))
+
+    if umap_version < LooseVersion('0.3.0'):
         from . import __version__
         # make this a warning, not an error
         # it might be useful for people to still be able to run it
         logg.warning(
             f'Scanpy {__version__} needs umap '
-            f'version >=0.3.0, not {umap.__version__}.'
+            f'version >=0.3.0, not {umap_version}.'
         )