--- a/scanpy/utils.py
+++ b/scanpy/utils.py
@@ -25,7 +25,12 @@ EPS = 1e-15
 
 def check_versions():
     from distutils.version import LooseVersion
-    from importlib_metadata import version
+    import importlib
+
+    if importlib.util.find_spec("importlib.metadata") is not None:
+        from importlib.metadata import version
+    else:
+        from importlib_metadata import version
 
     if sys.version_info < (3, 6):
         warnings.warn('Scanpy prefers Python 3.6 or higher. '