Follow up to #703, this should take import times down another second. It does add a dependency, but that dependency will be in the stdlib from 3.8+.

After this it looks like the next biggest contributors to import times are caused by `numba` (which would be hard to factor out) and `sklearn.metrics`, but these are on a smaller scale. Evidence of speedup:

```sh
isaac@Mimir ~/github/scanpy$ time python3 -c "import scanpy as sc"                      ✹ ✭master 
python3 -c "import scanpy as sc"  3.10s user 0.41s system 116% cpu 3.010 total
isaac@Mimir ~/github/scanpy$ git checkout defer-umap                                    ✹ ✭master 
Switched to branch 'defer-umap'
Your branch is up to date with 'origin/defer-umap'.
isaac@Mimir ~/github/scanpy$ time python3 -c "import scanpy as sc"                  ✹ ✭defer-umap 
python3 -c "import scanpy as sc"  2.18s user 0.42s system 131% cpu 1.981 total
```