@flying-sheep 2019-06-24
Please make the import conditional using

1. a `python_version < 3.8` [environment marker](https://www.python.org/dev/peps/pep-0345/#environment-markers) and
2. Try importing the 3.8 location (`importlib.metadata`) first before importing the backport module.