@ivirshup 2019-06-25
Done.

Just so I can figure out how I did it later, here's the commands for profiling import times:

```python
import cProfile
from pstats import SortKey, Stats

# This imports scanpy to current env, can't be run twice in a session
cProfile.run("import scanpy", "import-scanpy.profile")
p = Stats("import-scanpy.profile")
p.sort_stats(SortKey.CUMULATIVE)
p.print_stats(30)
```