@flying-sheep 2019-06-25
I started to test also with 3.8, but sadly there’s a bug in scipy: scipy/scipy#10354