@ivirshup 2019-06-25
I think the `<` should work, it's part of [PEP-508](https://www.python.org/dev/peps/pep-0508/#motivation). At least that's what [this blog post](https://hynek.me/articles/conditional-python-dependencies/) says.

Didn't those builds just fail before they got to that requirement?