@ivirshup 2019-06-27
I don't think we need to test against 3.8 until a stable release is out. I'm thinking we can just drop that from travis and merge?