@LuckyMD 2019-07-16
I'm getting an error loading scanpy (#739 ), and it points to the line you moved about deferring loading of umap-learn. 

When I revert back to commit abf95c645828e29edf5a7a27b05d9397f3c36f65 (the commit a couple before this), it works.