@flying-sheep 2019-07-16
Doesn’t seem like a problem with this. Further discussion in #739